Moved TimedTrigger into its own repository
Another separate of utility module so that later we can better control
what goes into each bundle when we separate wep and react code.
