+++ package.json
@@ -53,6 +53,7 @@
     "./packages/hyperion-async-counter",
     "./packages/hyperion-hook",
     "./packages/hyperion-channel",
+    "./packages/hyperion-timed-trigger",
     "./packages/hyperion-core",
     "./packages/hyperion-dom",
     "./packages/hyperion-flowlet",
@@ -67,6 +68,7 @@
     "@hyperion/hyperion-async-counter": "*",
     "@hyperion/hyperion-hook": "*",
     "@hyperion/hyperion-channel": "*",
+    "@hyperion/hyperion-timed-trigger": "*",
     "@hyperion/hyperion-autologging": "*",
     "@hyperion/hyperion-autologging-visualizer": "*",
     "@hyperion/hyperion-core": "*",
