+++ packages/hyperion-flowlet/src/FlowletManager.ts
@@ -7,7 +7,7 @@ import { InterceptableFunction, getFunctionInterceptor, interceptFunction } from
 import { CallbackType, interceptEventListener, isEventListenerObject } from "@hyperion/hyperion-dom/src/IEventListener";
 import { Flowlet } from "./Flowlet";
 import { assert, getLogger } from "@hyperion/hyperion-global";
-import { TimedTrigger } from "@hyperion/hyperion-util/src/TimedTrigger";
+import { TimedTrigger } from "@hyperion/hyperion-timed-trigger/src/TimedTrigger";
 
 const IS_FLOWLET_SETUP_PROP_NAME = `__isFlowletSetup`;
