+++ packages/hyperion-timed-trigger/.gitignore
@@ -1,0 +1,7 @@
+*.log
+.DS_Store
+node_modules
+dist
+coverage
+package-lock.json
+test/**/*.js
