+++ packages/hyperion-timed-trigger/babel.config.js
@@ -1,0 +1,2 @@
+const baseConfig = require('@hyperion/hyperion-devtools/babel.config');
+module.exports = baseConfig;
