+++ packages/hyperion-timed-trigger/jest.config.js
@@ -1,0 +1,6 @@
+const baseConfig = require('@hyperion/hyperion-devtools/jest.config');
+/*
+ * For a detailed explanation regarding each configuration property and type check, visit:
+ * https://jestjs.io/docs/configuration
+ */
+module.exports = baseConfig;
