+++ packages/hyperion-timed-trigger/package.json
@@ -1,0 +1,22 @@
+{
+  "scope": "@hyperion",
+  "name": "@hyperion/hyperion-timed-trigger",
+  "description": "Timed Trigger function to run one directly or based on a timer",
+  "version": "0.1.0",
+  "license": "MIT",
+  "scripts": {
+    "build": "tsc",
+    "watch": "npm run build -- --watch",
+    "start": "npm run watch",
+    "test": "npm run build & jest"
+  },
+  "workspaces": [
+    "../hyperion-devtools"
+  ],
+  "dependencies": {
+  },
+  "devDependencies": {
+    "@hyperion/hyperion-devtools": "*",
+    "@types/jest": "^29.5.3"
+  }
+}
