+++ packages/hyperion-timed-trigger/src/TimedTrigger.ts
