+++ packages/hyperion-timed-trigger/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  // see https://www.typescriptlang.org/tsconfig to better understand tsconfigs
+  "extends": "@hyperion/hyperion-devtools",
+  "include": [
+    "src",
+  ],
+}
