+++ src/index.ts
@@ -7,7 +7,12 @@ export { AsyncCounter } from "@hyperion/hyperion-async-counter/src/AsyncCounter"
 
 // hyperionHook
 export { Hook } from "@hyperion/hyperion-hook/src/Hook";
+
+// hyperionChannel
 export { PipeableEmitter, Channel } from "@hyperion/hyperion-channel/src/Channel";
+
+// hyperionTimedTrigger
+export { TimedTrigger } from "@hyperion/hyperion-timed-trigger/src/TimedTrigger";
 
 // hyperionCore
 export { setAssertLoggerOptions } from "@hyperion/hyperion-global/src/assert";
@@ -41,7 +46,6 @@ export * as SyncMutationObserver from "@hyperion/hyperion-util/src/SyncMutationO
 export { ClientSessionID } from "@hyperion/hyperion-util/src/ClientSessionID";
 export { SessionPersistentData, LocalStoragePersistentData } from "@hyperion/hyperion-util/src/PersistentData";
 export { default as TestAndSet } from "@hyperion/hyperion-util/src/TestAndSet";
-export { TimedTrigger } from "@hyperion/hyperion-util/src/TimedTrigger";
 
 // hyperionFlowletCore
 export { Flowlet, onFlowletInit } from "@hyperion/hyperion-flowlet/src/Flowlet";
